requirements.txt
# external requirements

joblib==1.0.1
numpy==1.21.1
pandas==1.3.0
pylint==2.10.0
PyYAML==6.0
scikit_learn==1.0.2
scipy==1.9.1
mlflow==1.19.0
boto3==1.26.28
matplotlib==3.6.2
fastapi==0.88.0
uvicorn[standard]==0.20.0
httpx==0.23.1
fastapi-utils==0.2.1
pytest==7.3.0
streamlit == 1.26.0
marshmallow_dataclass==8.6.0
python-dotenv==1.0.0
tqdm==4.65.0
protobuf==3.20.0
python-multipart==0.0.6
prometheus-fastapi-instrumentator==6.1.0
prometheus-client==0.17.1
black
flake8
isort
